Kinda stable right now. added buttons. fixed velocity to act as needed. planets behave different but still accurate.
