+++ PhysicSSimulator.vcxproj.filters
@@ -60,7 +60,7 @@
     <ClCompile Include="Serialization.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="PhysicsSimulationLocal.cpp">
+    <ClCompile Include="SinglePlayer.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="ElectricalParticle.cpp">
@@ -113,7 +113,7 @@
     <ClInclude Include="Serialization.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="PhysicsSimulation_Local.h">
+    <ClInclude Include="SinglePlayer.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="ElectricalParticle.h">
@@ -124,5 +124,22 @@
     <ResourceCompile Include="Engine.rc">
       <Filter>Resource Files</Filter>
     </ResourceCompile>
+  </ItemGroup>
+  <ItemGroup>
+    <Image Include="icon3.ico">
+      <Filter>Resource Files</Filter>
+    </Image>
+    <Image Include="icon1.ico">
+      <Filter>Resource Files</Filter>
+    </Image>
+    <Image Include="icon2.ico">
+      <Filter>Resource Files</Filter>
+    </Image>
+    <Image Include="Visuals\logo.png">
+      <Filter>Resource Files</Filter>
+    </Image>
+    <Image Include="Visuals\logo.ico">
+      <Filter>Resource Files</Filter>
+    </Image>
   </ItemGroup>
 </Project>
