+++ Planet.h
@@ -33,33 +33,6 @@ public:
 		return "Planet";
 	}
 
-	////For planet to planet interaction
-	//sf::Vector2f GravitateAccurate(BaseShape* object) {
-	//	// Calculate the vector from this object to the other object
-	//	sf::Vector2f distanceVec = GetPosition() - object->GetPosition();
-
-	//	// Calculate the distance between the two objects
-	//	float distanceLength = std::sqrt(distanceVec.x * distanceVec.x + distanceVec.y * distanceVec.y);
-
-	//	// Avoid division by zero or extremely small distances
-	//	const float epsilon = 1e-6f;
-	//	float distanceSquared = distanceLength* distanceLength;
-	//	if (distanceSquared < epsilon * epsilon || distanceSquared <= (4 * radius * radius + radius / 10)) {
-	//		return sf::Vector2f(0, 0);
-	//	}
-
-	//	// Normalize the distance vector to get the direction of the gravitational force
-	//	sf::Vector2f normalVector = distanceVec / distanceLength;
-
-	//	// Gravitational constant
-	//	const float G = 6.67430e-11f;
-
-	//	// Calculate the gravitational force magnitude using the formula:
-	//	// F = G * (m1 * m2) / d^2
-	//	float forceMagnitude = (G * GetMass() * object->GetMass()) / (distanceLength * distanceLength);
-	//	// return the force in the direction of the normalized vector
-	//	return normalVector * -forceMagnitude;
-	//}
 
 	sf::Vector2f GravitateAccurate(BaseShape* object) {
 		// Calculate the vector from this object to the other object
@@ -93,31 +66,31 @@ public:
 	}
 
 	//For planet and non planet interaction
-	void Gravitate(BaseShape* object)
+	void Gravitate(BaseShape* object, float dt)
 	{
-		// Calculate distance vector
-		float distance_x = object->GetPosition().x - GetPosition().x;
-		float distance_y = object->GetPosition().y - GetPosition().y;
+		// Calculate the vector from this object (Planet) to the other object
+		float distance_x = GetPosition().x - object->GetPosition().x;
+		float distance_y = GetPosition().y - object->GetPosition().y;
 
-		// Calculate total distance
+		// Calculate the distance between the two objects
 		float distance = sqrt(distance_x * distance_x + distance_y * distance_y);
 
-		// Prevent division by zero
-		if (distance < 1e-6f) return;
+		// Avoid division by zero in case the objects are at the same position. and if the object too close to the planet do not calculate so it wont fly away
+		if (distance == 0 || distance <= std::abs(object->GetEstimatedSize() * 2)) return;
 
-		// Inverse distance calculations
+		// Normalize the direction vector
 		float inverse_distance = 1.f / distance;
 		float normalized_x = inverse_distance * distance_x;
 		float normalized_y = inverse_distance * distance_y;
 
-		// Inverse square dropoff calculation
+		// Calculate gravitational force with inverse square law
 		float inverse_square_dropoff = inverse_distance * inverse_distance;
 
-		// Calculate acceleration components
+		// Calculate acceleration due to gravity based on innerGravity (strength of gravity)
 		float acceleration_x = normalized_x * innerGravity * inverse_square_dropoff;
 		float acceleration_y = normalized_y * innerGravity * inverse_square_dropoff;
 
-		// Accumulate gravitational acceleration
-		object->SetAcceleration(object->GetAcceleration() - sf::Vector2f(acceleration_x, acceleration_y));
+		object->SetAcceleration(sf::Vector2f(acceleration_x, acceleration_y));
 	}
+
 };
